refactor!: expose keys via env variable (#29)
the access and secret key are now registered as secret and then exposed
to s3cmd via the env variables AWS_ACCESS_KEY and AWS_SECRET_KEY. This
prevents the need to write them to disk. However, it can happen that
this interferes with other values the user needs for these variables.
Usually it should be aligned, even if these keys are used for other
purposes. Because they operate on the same account.

Signed-off-by: Nico Braun <rainbowstack@gmail.com>